% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{loan_limit}
\alias{loan_limit}
\title{What is the possible limit?}
\usage{
loan_limit(rawfile)
}
\arguments{
\item{rawfile}{MPESA pdf file formated by mpesa_file() function}
}
\value{
\code{statement_period(rawfile)} .
}
\description{
Gives you the posible limit a client should be granted.The limit is calculated by getting 75% of the monthly average Credit ie "PaidIN" in this case
}
\examples{
library(mpesaR)
limit <- loan_limit(rawfile)
}
