% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{mpesa_file}
\alias{mpesa_file}
\title{Imports and formats MPESA file.}
\usage{
mpesa_file(path, password = NULL)
}
\arguments{
\item{path}{Path to your file.}

\item{password}{MPESA statement password.A pop up will appear if not filled in.hint your national id}
}
\description{
This function loads and formats the raw pdf statement int your work space.Its important to note that this function is built over pdftools package.The function thus adds more functionality to pdftools::pdf_text() function.
}
\examples{
mpesa_file(path = "path/to/my/mpesa_statement.pdf",password = "my secret password")
}
