% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{statement_full}
\alias{statement_full}
\title{Get Full Statement}
\usage{
statement_full(rawfile)
}
\arguments{
\item{rawfile}{MPESA pdf file formated by mpesa_file() function}
}
\value{
\code{statement_full(rawfile)} returns tidy dataframe of the MPESA statements tables.
}
\description{
This function extracts all mpesa statement tables for all pages ,cleans and presents it as a tidy dataframe.
}
\examples{
library(mpesaR)
statement_full(rawfile)
#using the pipe ( \%>\% ) operator
mpesa_file("path/to/MPESAfile.pdf","my_national_id") \%>\%
statement_full()
}
