% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{statement_header}
\alias{statement_header}
\title{Get statement Bio Data}
\usage{
statement_header(rawfile)
}
\arguments{
\item{rawfile}{MPESA pdf file formated by mpesa_file() function}
}
\value{
\code{statement_header(rawfile)} returns tidy dataframe of  MPESA statement header information
}
\description{
This function enables you to extract biodata information from an MPESA statement.
}
\examples{
library(mpesaR)
rawfile = mpesa_file()
statement_header(rawfile)
#using pipe operator
mpesa_file("path/to/MPESAfile.pdf","my_national_id") \%>\%
statement_header()
}
