% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{statement_in}
\alias{statement_in}
\title{Where did money come from?}
\usage{
statement_in(statement)
}
\arguments{
\item{statement}{This is the full MPESA statement table generated by statement_full() function.Please note that this function does not work without the statement.}
}
\value{
\code{statement_in(statement)} returns a dataframe with sources of income
}
\description{
Quckly generate extract vital information on the sources of icoming cash to the statement.
}
\examples{
library(mpesaR)
statement <- statement_full(rawfile)
statement_in(statement)
#using the pipe ( \%>\% ) operator
mpesa_file("path/to/MPESAfile.pdf","MYSECRETPASWWORD") \%>\%
statement_full() \%>\%
statement_in()
}
