% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{statement_kpis}
\alias{statement_kpis}
\title{Key KPIS}
\usage{
statement_kpis(statement)
}
\arguments{
\item{statement}{This is the full MPESA statement table generated by statement_full() function.Please note that this function does not work without the statement.}
}
\value{
\code{statement_kpis(statement)} returns key kpis from the mpesa statement
}
\description{
Quckly generate key kpis from the MPESA statement supplied.
}
\examples{
library(mpesaR)
statement <- statement_full(rawfile)
statement_kpis(statement)
#using the pipe ( \%>\% ) operator
mpesa_file("path/to/MPESAfile.pdf","MYSECRETPASWWORD") \%>\%
statement_full() \%>\%
statement_kpis()
}
