% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{statement_out}
\alias{statement_out}
\title{Where did expense come from?}
\usage{
statement_out(statement)

statement_out(statement)
}
\arguments{
\item{statement}{This is the full MPESA statement table generated by statement_full() function.Please note that this function does not work without the statement.}
}
\value{
\code{statement_out(statement)} returns a dataframe with a list of sources of expenses

\code{statement_out(statement)} returns a dataframe with a list of sources of expenses
}
\description{
Quckly generate and extract vital information on where most expenses come from.

Quckly generate and extract vital information on where most expenses come from.
}
\examples{
library(mpesaR)
statement <- statement_full(rawfile)
statement_out(statement)
#using the pipe ( \%>\% ) operator
mpesa_file("path/to/MPESAfile.pdf","MYSECRETPASWWORD") \%>\%
statement_full() \%>\%
statement_out()
library(mpesaR)
statement <- statement_full(rawfile)
statement_out(statement)
#using the pipe ( \%>\% ) operator
mpesa_file("path/to/MPESAfile.pdf","MYSECRETPASWWORD") \%>\%
statement_full() \%>\%
statement_out()
}
