% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{statement_period}
\alias{statement_period}
\title{What is the transaction period?}
\usage{
statement_period(rawfile)
}
\arguments{
\item{rawfile}{MPESA pdf file formated by mpesa_file() function}
}
\value{
\code{statement_period(rawfile)} returns period difference
}
\description{
Gives you the transaction period from the start of the statement?.
}
\examples{
library(mpesaR)
statement <- statement_full(rawfile)
statement_period(statement)
#using the pipe ( \%>\% ) operator
mpesa_file("path/to/MPESAfile.pdf","MYSECRETPASWWORD") \%>\%
statement_full() \%>\%
statement_period()
}
