% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{statement_summary}
\alias{statement_summary}
\title{Get Statement Summary statistics}
\usage{
statement_summary(rawfile)
}
\arguments{
\item{rawfile}{MPESA pdf file formated by mpesa_file() function}
}
\value{
\code{statement_summary(rawfile)} returns tidy dataframe of the MPESA summary statistics
}
\description{
This function extracts the statement summary at the begining of every mpesa statement file
}
\examples{
library(mpesaR)
statement_summary(rawfile)
#using the pipe ( \%>\% ) operator
mpesa_file("path/to/MPESAfile.pdf","my_national_id") \%>\%
statement_summary()
}
