% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpesa_file.R
\name{trx_no}
\alias{trx_no}
\title{How many transactions?}
\usage{
trx_no(statement)
}
\arguments{
\item{statement}{This is the full MPESA statement table generated by statement_full() function.Please note that this function does not work without the statement.}
}
\value{
\code{trx_no(rawfile)} returns a numeric value of the number of transactions.
}
\description{
This function gives you to get the number of transactions in a given financial reporting period
}
\note{
The function doesnt count the mpesa charge cases.
}
\examples{
library(mpesaR)
statement <- statement_full(rawfile)
trx_no(statement)
#using the pipe ( \%>\% ) operator
mpesa_file("path/to/MPESAfile.pdf","my_national_id") \%>\%
statement_full() \%>\%
trx_no()
}
